utils: move nb_days_in_year to utils
Since geiq_assessments/sync.py imports geiq_assessments/models.py, using
_nb_days_in_year function in models.py would have caused a cycle.

It makes more sense to move it to utils where it might be used for other
parts.
